+++ features/auth/signup/components/signup-flow.tsx
@@ -6,12 +6,12 @@ import { FormProvider } from 'react-hook-form';
 
 import { StepIndicator } from '@/shared/components/step-indicator';
 
-import type { SignupStepId } from '../signup-constants';
+import type { SignupStepId } from '../signup.constants';
 import type { SignupFormData } from '../types';
 
 import { SignupStepCareer } from './signup-step-career';
 import { SignupStepComplete } from './signup-step-complete';
-import { SIGNUP_STEPS, SIGNUP_STEP_DESCRIPTIONS } from '../signup-constants';
+import { SIGNUP_STEPS, SIGNUP_STEP_DESCRIPTIONS } from '../signup.constants';
 import { SignupStepBasic } from './signup-step-basic';
 import { useSignupForm } from '../hooks/use-signup-form';
