+++ features/auth/signup/hooks/use-agreements.ts
@@ -3,7 +3,7 @@ import { useWatch } from 'react-hook-form';
 import type { SignupFormData } from '../types';
 import type { Control, UseFormSetValue } from 'react-hook-form';
 
-import { AGREEMENTS } from '../signup-constants';
+import { AGREEMENTS } from '../signup.constants';
 
 const defaultAgreements: SignupFormData['agreements'] = {
   age: false,
