+++ features/auth/signup/hooks/use-signup.mutation.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { profileKeys } from '@/features/profile/profile.query-keys';
 
-import { createSignup } from '../create-signup.api';
+import { createSignup } from '../api/create-signup.api';
 
 export const useSignup = () => {
   const queryClient = useQueryClient();
