+++ features/auth/signup/types.ts
@@ -1,5 +1,5 @@
-import type { FIELD_VALUES, USER_LINK_TYPE_VALUES } from './signup-constants';
-import type { SignupFormSchema } from './signup-form.schema';
+import type { FIELD_VALUES, USER_LINK_TYPE_VALUES } from './signup.constants';
+import type { SignupFormSchema } from './signup.schemas';
 import type { z } from 'zod';
 
 export type SignupFormData = z.infer<typeof SignupFormSchema>;
